requirements.txt
fastapi~=0.115.12
uvicorn
python-dotenv~=1.1.0
starlette~=0.46.2
pillow~=11.0.0
selenium~=4.33.0
torch~=2.6.0
torchvision~=0.22.0
pydantic-settings~=2.9.1
transformers~=4.52.4
pydantic~=2.11.5